pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.13",
    "python-dotenv>=1.1.1",
    "python-telegram-bot==20.8",
    "telegram>=0.0.1",
]
